Renamed movesTable as sanMovesTable
