+++ src/styles/sanMovesTable.js
